Resolve clang-tidy findings across src/ and encapsulate domain members
Fix the genuine code-level findings and tighten encapsulation so static
analysis passes cleanly:

- Encapsulate previously public/protected members behind accessors:
  EventBus signals, panel/store signals, PageSetupStore data, the
  calendar-grid widgets (via friend), and the date-store re-entry flag.
- Name magic numbers: UTF-8 lead/continuation masks in font.hpp (which
  also documents the decode scheme) and GUI layout constants in the
  page/font/calendar panels.
- Portability and correctness: fix a most-vexing-parse in
  value_serialization (clang rejected the inline date_period(date(...))
  form), drop the unused SceneNode::is_dirty field, replace a broken
  multi-line NOLINT with a real cast, avoid unnecessary string/vector
  copies, use std::ranges algorithms, and make float/int conversions
  explicit.

Behaviour is unchanged: the 200-dpi PNG render is byte-identical and all
41 unit tests pass.

+++ src/gui/shape_panel.hpp
@@ -263,7 +263,8 @@ class ElementsSetupsPanel : public wxPanel {
     fill_color_picker->Enable(fill_visible);
     fill_color_alpha_slider->Enable(fill_visible);
 
-    linewidth_ctrl->SetValue(current_configuration.LineWidthDisabled());
+    linewidth_ctrl->SetValue(
+        static_cast<double>(current_configuration.LineWidthDisabled()));
 
     constexpr float kAlphaScale = 100.0F;
     const auto outline_color = current_configuration.OutlineColorDisabled();
